requirements.txt
pandas
plotly
scrapy
geopy
streamlit
numpy
folium
streamlit_folium
requests
sqlalchemy
psycopg2
python-telegram-bot
